fix arrancar y detener (con permisos, cambian los numeros)

+++ AFRA-J/Detener.sh
@@ -1,8 +1,22 @@
 #!/bin/bash
+cant=$(ps aux | grep -c " .*$1");
 
-for word in $(ps | grep " .*$1")
+if [[ $cant < 4 ]]
+then
+	echo "No hay un proceso $1 corriendo";
+	exit 0
+fi
+
+CONT=1;
+for word in $(ps aux | grep " .*$1")
 do
-    kill $word
-    exit 0
+	if [[ $CONT == 1 ]]
+    then
+    	#ignoro el usuario
+    	CONT=$((CONT+1));
+    else
+    	kill $word
+    	exit 0
+    fi
 done
